Abandon null-terminated C strings for Strings
In golfscript, it's perfectly valid to have a null character in the middle of
a string (such as in the program [0]""+), and such strings were handled
incorrectly in the previous version of the interpreter, which I have now
hopefully fixed. I've tested it, and there doesn't seem to be any new bugs
introduced from this change, but I wouldn't be surprised if one slipped past
my radar.