Fix behavior of arrays of integers and strings
While testing things in the ruby golfscript interpreter, I discovered
that in a lot of contexts arrays of integers and strings are considered
to be the same (for instance, the program '[[97 98 99]] "abc" - p' will
output an empty array). Previously, I did not realize this was the case
and so this commit fixes that.