Fix redefining a function inside itself
Previously, if you reassigned a block inside itself, (such as in the
program "{{]{+}*}:a;1 1a}:a; a", some very unexpected things would
occur because the new definition would overwrite the old while still
using the old definition. This copies defined functions before they're
executed so we don't have that problem.