Implement arbitrary-length integers.
Originally, I decided to use fixed-width integers for the interpreter,
but I changed my mind and decided to implement arbitrary-length
integers. Anyway, this commit adds a new Bigint struct, and replaces
int64_t with Bigint in the Item struct, and replaces a lot of code with
functions operating on Bigints.

There are definitely areas that can be improved, and there are likely
to be bugs that I haven't found, so there are likely to be many commits
to follow to improve Bigints.