Perform addition/subtraction/increment in place
Adding to or subtracting from a number used to require the creation of
a temporary Bigint, which made the code look unclean and also was just
unnecessary, so I changed bigint_add and bigint_subtract so they store
the calculated value in the first parameter, rather than returning the
result.

This also changes bigint_increment and bigint_decrement to get rid of
the need for the creation of a temporary Bigint just for the increment
or decrement.